requirements.txt
# Core Django
django>=5.0
djangorestframework
django-cors-headers
python-dotenv
gunicorn>=21.2.0
djangorestframework-simplejwt>=5.3.0

# Database
mysql-connector-python>=8.0

# File handling
pillow
django-storages

# Filtering
django-filter

# Background Tasks & Scheduling
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Timezone handling
pytz

# Email (SMTP - built into Django)
# No additional package needed
celery


# WhatsApp Business API Client
# Using official WhatsApp Business Cloud API via HTTP requests
requests>=2.31.0

# Google Sheets Integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-decouple>=3.8